# Optimize _expand_named_fields