@wimglenn 2023-11-09
Seems reasonable. This is not exactly logically equivalent, for example if the input was `'[aaa]'` the existing code will raise, but this code will return basename, subkeys as `""`, `"[aaa]"`. But I don't see any handling around the `AttributeError` so I don't think it should cause any issue.